# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Documentation API",
        "description": "API untuk integrasi dengan aplikasi Workshop. Endpoint ini dipakai aplikasi terpisah untuk menerima Permintaan Unit yang sudah diteruskan, menyetujui penyiapan unit, dan mengupdate kesiapan tiap item.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "SSB Project API Server"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Generate Bearer Token (Sanctum) — sementara",
                "description": "Endpoint sementara untuk testing. Login dengan email & password user yang sudah ada, dapatkan Bearer Token untuk dipakai di tombol Authorize di atas.",
                "operationId": "2ed45531e4e8326a9e2713d5925fb861",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@ssb.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil — token diterbitkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef1234567890..."
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Admin"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@ssb.test"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/unit-requests": {
            "get": {
                "tags": [
                    "Unit Request"
                ],
                "summary": "List Permintaan Unit yang sudah diteruskan ke Workshop",
                "description": "Default-nya hanya menampilkan status FORWARDED_TO_WORKSHOP. Gunakan parameter ?status=APPROVED_FROM_WORKSHOP, ?status=REJECTED_FROM_WORKSHOP, kombinasi (comma-separated), atau ?status=all untuk mengubah filter.",
                "operationId": "d8dbfc38a85629f8ef9821902c9feb2a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter status. Default: FORWARDED_TO_WORKSHOP. Boleh comma-separated (mis. FORWARDED_TO_WORKSHOP,APPROVED_FROM_WORKSHOP) atau 'all'.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "FORWARDED_TO_WORKSHOP"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per halaman (1-100). Default: 25.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 25,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Permintaan Unit (paginated)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Parameter status tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/unit-requests/{uid}": {
            "get": {
                "tags": [
                    "Unit Request"
                ],
                "summary": "Detail satu Permintaan Unit",
                "operationId": "3124d7907444dc5a4886bf219792503f",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "UID (UUID) Permintaan Unit",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail Permintaan Unit"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/unit-requests/{uid}/status": {
            "patch": {
                "tags": [
                    "Unit Request"
                ],
                "summary": "Update status Permintaan Unit",
                "description": "Workshop memakai endpoint ini untuk menyetujui atau menolak Permintaan Unit yang diteruskan. Hanya transisi yang diizinkan UnitRequestStatus::canTransitionTo() yang diterima (mis. FORWARDED_TO_WORKSHOP → APPROVED_FROM_WORKSHOP / REJECTED_FROM_WORKSHOP).",
                "operationId": "9be12970a4f84e3227bf2a557f3bfe3d",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "UID (UUID) Permintaan Unit",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Status tujuan. Untuk Workshop: APPROVED_FROM_WORKSHOP (menyetujui penyiapan unit) atau REJECTED_FROM_WORKSHOP (menolak permintaan).",
                                        "type": "string",
                                        "enum": [
                                            "DRAFT",
                                            "SUBMITTED",
                                            "APPROVED",
                                            "REJECTED",
                                            "FORWARDED_TO_WORKSHOP",
                                            "APPROVED_FROM_WORKSHOP",
                                            "REJECTED_FROM_WORKSHOP"
                                        ],
                                        "example": "APPROVED_FROM_WORKSHOP"
                                    },
                                    "notes": {
                                        "description": "Catatan opsional dari Workshop (mis. alasan penolakan).",
                                        "type": "string",
                                        "example": "Unit siap kirim per 2026-05-25",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status berhasil diupdate"
                    },
                    "422": {
                        "description": "Status tidak valid atau transisi tidak diizinkan"
                    },
                    "404": {
                        "description": "Permintaan Unit tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}